fix: fix digit loss of sweep frequency

+++ nanovna.h
@@ -71,7 +71,7 @@ enum {
   ST_START, ST_STOP, ST_CENTER, ST_SPAN, ST_CW
 };
 
-void set_sweep_frequency(int type, float frequency);
+void set_sweep_frequency(int type, uint32_t frequency);
 uint32_t get_sweep_frequency(int type);
 
 float my_atof(const char *p);
